+++ browser_library_test/rollup.config.require.mjs
@@ -1,0 +1,11 @@
+import { nodeResolve } from '@rollup/plugin-node-resolve';
+import commonjs from '@rollup/plugin-commonjs';
+
+export default {
+  input: 'lib/require.js',
+  output: {
+    file: 'lib/build/rollup-require.js',
+format: 'iife',
+  },
+  plugins: [nodeResolve({browser: true}), commonjs()],
+};